requirements.txt
# ABOUTME: Minimal dependencies for direct mesh-to-gaussian conversion
# ABOUTME: PyTorch is optional - only needed for GPU optimization

# Core dependencies
numpy>=1.24.0
scipy>=1.11.0
trimesh>=3.23.0
pillow>=10.0.0
psutil>=5.9.0

# Viewer dependencies (for gaussian-splat-viewer)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
watchdog>=3.0.0
python-multipart>=0.0.6

# Optional GPU optimization
# Uncomment to enable:
# torch>=2.0.0
# torchvision>=0.15.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

